unified-debug-agent/requirements.txt
# Core dependencies
langchain==0.3.7
langchain-openai==0.2.8
langchain-mcp-adapters==0.1.2
pydantic==2.10.2
python-dotenv==1.0.1

# Vector similarity and embeddings
qdrant-client==1.12.1
sentence-transformers==3.2.1

# Code analysis
tree-sitter==0.23.2
tree-sitter-python==0.23.2
tree-sitter-javascript==0.23.0
tree-sitter-java==0.23.2
tree-sitter-cpp==0.23.0
tree-sitter-go==0.23.1

# API clients
httpx==0.28.1
requests==2.32.3

# Utilities
GitPython==3.1.43
pathspec==0.12.1